feat: moved types from sdk project into runtime
Since we are now requiring the runtime as an explicit dependency for Node.js Function projects then we no longer need the overhead of maintaining separate SDK type definitions in `@heroku/sf-fx-sdk-nodejs`.

[GUS-W-11944771](https://gus.lightning.force.com/lightning/r/ADM_Work__c/a07EE000019Rqb8YAC/view)
