+++ src/sdk/user.ts
@@ -5,7 +5,7 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import { User } from "../types";
+import { User } from "../index";
 
 export class UserImpl implements User {
   readonly id: string;
